Added headers comments and netscape parameters in some file

+++ src/test/fork.t.sol
@@ -20,13 +20,15 @@ pragma solidity ^0.6.12;
 import "ds-test/test.sol";
 import "ds-token/token.sol";
 
-import {Vat} from '../vat.sol';
+import {Vat} from "../vat.sol";
 
 contract Usr {
     Vat public vat;
+
     constructor(Vat vat_) public {
         vat = vat_;
     }
+
     function try_call(address addr, bytes calldata data) external returns (bool) {
         bytes memory _data = data;
         assembly {
@@ -37,7 +39,8 @@ contract Usr {
             revert(free, 32)
         }
     }
-    function can_frob(bytes32 ilk, address u, address v, address w, int dink, int dart) public returns (bool) {
+
+    function can_frob(bytes32 ilk, address u, address v, address w, int256 dink, int256 dart) public returns (bool) {
         string memory sig = "frob(bytes32,address,address,address,int256,int256)";
         bytes memory data = abi.encodeWithSignature(sig, ilk, u, v, w, dink, dart);
 
@@ -47,7 +50,8 @@ contract Usr {
         ok = abi.decode(success, (bool));
         if (ok) return true;
     }
-    function can_fork(bytes32 ilk, address src, address dst, int dink, int dart) public returns (bool) {
+
+    function can_fork(bytes32 ilk, address src, address dst, int256 dink, int256 dart) public returns (bool) {
         string memory sig = "fork(bytes32,address,address,int256,int256)";
         bytes memory data = abi.encodeWithSignature(sig, ilk, src, dst, dink, dart);
 
@@ -57,12 +61,15 @@ contract Usr {
         ok = abi.decode(success, (bool));
         if (ok) return true;
     }
-    function frob(bytes32 ilk, address u, address v, address w, int dink, int dart) public {
+
+    function frob(bytes32 ilk, address u, address v, address w, int256 dink, int256 dart) public {
         vat.frob(ilk, u, v, w, dink, dart);
     }
-    function fork(bytes32 ilk, address src, address dst, int dink, int dart) public {
+
+    function fork(bytes32 ilk, address src, address dst, int256 dink, int256 dart) public {
         vat.fork(ilk, src, dst, dink, dart);
     }
+
     function hope(address usr) public {
         vat.hope(usr);
     }
@@ -76,10 +83,11 @@ contract ForkTest is DSTest {
     address a;
     address b;
 
-    function ray(uint wad) internal pure returns (uint) {
+    function ray(uint256 wad) internal pure returns (uint256) {
         return wad * 10 ** 9;
     }
-    function rad(uint wad) internal pure returns (uint) {
+
+    function rad(uint256 wad) internal pure returns (uint256) {
         return wad * 10 ** 27;
     }
 
@@ -91,37 +99,41 @@ contract ForkTest is DSTest {
         b = address(bob);
 
         vat.init("gems");
-        vat.file("gems", "spot", ray(0.5  ether));
+        vat.file("gems", "spot", ray(0.5 ether));
         vat.file("gems", "line", rad(1000 ether));
-        vat.file("Line",         rad(1000 ether));
+        vat.file("Line", rad(1000 ether));
 
         vat.slip("gems", a, 8 ether);
     }
+
     function test_fork_to_self() public {
         ali.frob("gems", a, a, a, 8 ether, 4 ether);
-        assertTrue( ali.can_fork("gems", a, a, 8 ether, 4 ether));
-        assertTrue( ali.can_fork("gems", a, a, 4 ether, 2 ether));
+        assertTrue(ali.can_fork("gems", a, a, 8 ether, 4 ether));
+        assertTrue(ali.can_fork("gems", a, a, 4 ether, 2 ether));
         assertTrue(!ali.can_fork("gems", a, a, 9 ether, 4 ether));
     }
+
     function test_give_to_other() public {
         ali.frob("gems", a, a, a, 8 ether, 4 ether);
         assertTrue(!ali.can_fork("gems", a, b, 8 ether, 4 ether));
         bob.hope(address(ali));
-        assertTrue( ali.can_fork("gems", a, b, 8 ether, 4 ether));
+        assertTrue(ali.can_fork("gems", a, b, 8 ether, 4 ether));
     }
+
     function test_fork_to_other() public {
         ali.frob("gems", a, a, a, 8 ether, 4 ether);
         bob.hope(address(ali));
-        assertTrue( ali.can_fork("gems", a, b, 4 ether, 2 ether));
+        assertTrue(ali.can_fork("gems", a, b, 4 ether, 2 ether));
         assertTrue(!ali.can_fork("gems", a, b, 4 ether, 3 ether));
         assertTrue(!ali.can_fork("gems", a, b, 4 ether, 1 ether));
     }
+
     function test_fork_dust() public {
         ali.frob("gems", a, a, a, 8 ether, 4 ether);
         bob.hope(address(ali));
-        assertTrue( ali.can_fork("gems", a, b, 4 ether, 2 ether));
+        assertTrue(ali.can_fork("gems", a, b, 4 ether, 2 ether));
         vat.file("gems", "dust", rad(1 ether));
-        assertTrue( ali.can_fork("gems", a, b, 2 ether, 1 ether));
+        assertTrue(ali.can_fork("gems", a, b, 2 ether, 1 ether));
         assertTrue(!ali.can_fork("gems", a, b, 1 ether, 0.5 ether));
     }
 }
